The following function textually and audibly communicates a phrase.
The following function textually and audibly communicates a phrase. 
The open source eSpeak speech synthesizer is used to create the audible message. 
If the eSpeak exe is not located in the directory, no audible message will be heard.
Concat the phrase to the command
Convert to a const char
Textually output phrase
Send the command to cmd to execute espeak with the phrase an argument