The following function will look for the passed phrase in the memory file.
The following function will look for the passed phrase in the memory file. If there is a match, the accompanying response, stored below the initial phrase will be outputted. If the response cannot be found, the learner will repeat the phrase, and prompt the user to enter an ideal response. This response will be stored in the memory file along with the initial phase.